[camp-lua] Split user object factory ("new" function) and Lua metaclass
"context["A"] = camp::classByType<A>()" will now affect a "factory object" (a Lua table) to the A
global. The factory has a "new" function which will construct and return a new A instance. The
metaclass of the newly created instance contains only the "__index" and "__newindex" events. Thus,
it is now possible to have a "new" camp property or function.